Add : 컨텐츠 작성 시 이미지 배열 업로드

+++ controller/content.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { Iuser } from '@interface';
+import { Iuser, Icontent, Iartist, IcontentImages } from '@interface';
 import { v4 as uuidv4 } from 'uuid';
 import { findArtistById } from '@database/artists';
 import {
@@ -9,7 +9,8 @@ import {
   updateContent,
   deleteContent,
 } from '@database/contents';
-import { findUserById } from '@database/users';
+import { uploadImage } from '@util/aws';
+import multer from 'multer';
 
 export const createContents = async (
   req: Request,
@@ -17,31 +18,46 @@ export const createContents = async (
 ): Promise<void> => {
   try {
     const { tokenUser } = req;
+    const { images: imageData } = req.files as IcontentImages;
     const { id, name, profileImage } = tokenUser as Iuser;
+    console.log('imageData : ', imageData);
     const {
       artistId,
       title,
       tags,
       description,
-      images,
       date,
       time,
       address,
       mobile,
       perks,
     } = req.body;
-
+    const imageUrls = [];
     const celeb = await findArtistById(artistId);
+    if (imageData) {
+      for (let idx = 0; idx < imageData.length; idx++) {
+        const data = await uploadImage(imageData[idx]);
+        imageUrls.push(data);
+      }
+    }
+    //text값으로 들어온 json 값을 parsing을 해서
+    const jsonTags = JSON.parse(tags);
+    const jsonDate = JSON.parse(date);
+    const jsonTime = JSON.parse(time);
+    const jsonAddress = JSON.parse(address);
+    const jsonPerks = JSON.parse(perks);
+    //tag date time address perks
+    //json.parse 해서 사용해야함
 
     if (!celeb) {
-      res.status(401).send({
-        message: 'unauthorized',
+      res.status(400).send({
+        message: 'no data',
       });
       return;
     }
-    const contentsId = uuidv4();
+    console.log('입력 값 : ', req.body);
     const newContent = await createContent({
-      id: contentsId,
+      id: uuidv4(),
       author: {
         id: id,
         name: name,
@@ -54,36 +70,21 @@ export const createContents = async (
         profileImage: celeb.profileImage,
       },
       title: title,
-      images: images,
-      date: {
-        start: date.start,
-        end: date.end,
-      },
-      time: {
-        open: time.open,
-        close: time.close,
-      },
-      address: {
-        storeName: address.storeName,
-        roadAddress: address.roadAddress,
-        location: {
-          lat: address.location.lat,
-          lng: address.location.lng,
-        },
-      },
+      images: imageUrls as string[],
+      date: jsonDate,
+      time: jsonTime,
+      address: jsonAddress,
       mobile: mobile,
       description: description,
-      tags: tags,
-      perks: perks,
+      tags: jsonTags,
+      perks: jsonPerks,
     });
 
     if (newContent) {
       res.status(201).send({ result: newContent, message: 'ok' });
-    } else {
-      res.status(400).send({ message: 'invalid input' });
     }
   } catch (err) {
-    console.error('create content error');
+    console.error('create content error', err.message);
     res.status(400).send({
       message: 'invalid request',
     });
@@ -97,8 +98,24 @@ export const deleteContents = async (
   try {
     const { tokenUser } = req;
     const { id } = req.body;
+    const author = (await findContentById(id)) as Icontent;
+    if (!author) {
+      res.status(404).send({
+        message: 'invalid request',
+      });
+      return;
+    }
+
+    const user = tokenUser as Iuser;
+    if (author.author.id !== user.id) {
+      res.status(401).send({
+        message: 'unauthorized',
+      });
+      return;
+    }
+
     const deleteResult = await deleteContent(id);
-    if (deleteResult && tokenUser) {
+    if (deleteResult) {
       res.status(201).send({ message: 'ok' });
     }
   } catch (err) {
@@ -128,15 +145,15 @@ export const updateContents = async (
       mobile,
       perks,
     } = req.body;
-
-    const celeb = await findArtistById(artistId);
-
-    if (!celeb) {
+    const user = tokenUser as Iuser;
+    const author = await findContentById(id);
+    if (user.id !== author?.author.id) {
       res.status(401).send({
         message: 'unauthorized',
       });
       return;
     }
+    const celeb = (await findArtistById(artistId)) as Iartist;
     const updateResult = await updateContent({
       id: id,
       title: title,
@@ -168,10 +185,7 @@ export const updateContents = async (
       tags: tags,
       perks: perks,
     });
-    if (!updateResult) {
-      res.status(400).send('invalid request');
-      return;
-    } else {
+    if (updateResult) {
       res.status(201).send({ result: updateResult, message: 'ok' });
     }
   } catch (err) {
@@ -220,6 +234,90 @@ export const listOfContents = async (req: Request, res: Response) => {
     res.status(200).send({ result: resultList, message: 'ok' });
   } catch (err) {
     console.error('contents list error', err.message);
+    res.status(400).send({
+      message: 'invalid request',
+    });
+  }
+};
+
+export const createContentsTesting = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
+  try {
+    const { tokenUser } = req;
+    const { id, name, profileImage } = tokenUser as Iuser;
+    const {
+      artistId,
+      title,
+      tags,
+      description,
+      date,
+      time,
+      address,
+      mobile,
+      perks,
+    } = req.body;
+    const imagesFiles = req.files as Express.Multer.File[];
+    console.log(imagesFiles);
+    const celeb = await findArtistById(artistId);
+    console.log(imagesFiles);
+    if (!celeb) {
+      res.status(400).send({
+        message: 'no data',
+      });
+      return;
+    }
+
+    const urls = imagesFiles.map((imagesFiles) => {
+      const url = uploadImage(imagesFiles);
+      return url;
+    });
+    console.log(urls);
+
+    const newContent = await createContent({
+      id: uuidv4(),
+      author: {
+        id: id,
+        name: name,
+        profileImage: profileImage,
+      },
+      artist: {
+        id: celeb.id,
+        name: celeb.name,
+        group: celeb.group,
+        profileImage: celeb.profileImage,
+      },
+      title: title,
+      //images: images,
+      images: [''],
+      date: {
+        start: date.start,
+        end: date.end,
+      },
+      time: {
+        open: time.open,
+        close: time.close,
+      },
+      address: {
+        storeName: address.storeName,
+        roadAddress: address.roadAddress,
+        location: {
+          lat: address.location.lat,
+          lng: address.location.lng,
+        },
+      },
+      mobile: mobile,
+      description: description,
+      tags: tags,
+      perks: perks,
+    });
+
+    if (newContent) {
+      res.status(201).send({ result: newContent, message: 'ok' });
+    }
+  } catch (err) {
+    console.error('create content error');
     res.status(400).send({
       message: 'invalid request',
     });
